Add files via upload
This is a Java program that prompts the user to convert an IP address to a DNS name. It uses the InetAddress class in the java.net package to get the IP address of the local host and convert it to a DNS name.

The program first prompts the user to see if they want to convert an IP address to a DNS name. If the user answers "y," the program gets the local host's IP address and prompts the user to see if they want to convert it to a DNS name. If the user answers "y" again, the program prompts the user to enter a name for the DNS, and then displays that name and prompts the user to test the DNS by entering the name they just entered. If the user enters the correct name, the program displays the IP address of the local host, and if they enter the IP address, the program displays the DNS name they just entered. If the user enters anything else, the program displays an error message.

If the user does not want to convert an IP address to a DNS name, the program displays a message saying that it can only convert IP addresses to DNS names.

Overall, the program is a simple command-line tool for converting IP addresses to DNS names, and it demonstrates the use of the InetAddress class in Java.